CV filltering/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
google-generativeai==0.3.2
supabase==1.2.0
requests==2.31.0
